compute/builtin: start new tests for node connection
