+++ compute/builtin/testutils_test.go
@@ -2,62 +2,53 @@ package builtin
 
 import (
 	"context"
+	"fmt"
 	"io/ioutil"
+	"net"
 	"testing"
 	"time"
 
 	"github.com/ohsu-comp-bio/funnel/config"
+	"github.com/ohsu-comp-bio/funnel/config/testconfig"
+	"github.com/ohsu-comp-bio/funnel/events"
 	"github.com/ohsu-comp-bio/funnel/logger"
-	"github.com/ohsu-comp-bio/funnel/util"
-	"github.com/stretchr/testify/mock"
+	"github.com/ohsu-comp-bio/funnel/util/rpc"
+	"google.golang.org/grpc"
 )
 
-// testNode wraps Node with some testing helpers.
-type testNode struct {
-	*NodeProcess
-	Client *MockClient
-	done   chan struct{}
+func testConfig() config.Config {
+	conf := config.DefaultConfig()
+	conf = testconfig.TestifyConfig(conf)
+	conf.Compute = "builtin"
+	conf.Node.UpdateRate = config.Duration(50 * time.Millisecond)
+	conf.Scheduler.NodePingTimeout = config.Duration(time.Millisecond * 300)
+	conf.Scheduler.NodeInitTimeout = config.Duration(time.Millisecond * 300)
+	conf.Scheduler.NodeDeadTimeout = config.Duration(time.Millisecond * 300)
+
+	workDir, err := ioutil.TempDir("", "funnel-test-node-")
+	if err != nil {
+		panic(err)
+	}
+
+	conf.Worker.WorkDir = workDir
+	conf.Node.WorkDir = workDir
+	return conf
 }
 
-func newTestNode(conf config.Config, t *testing.T) testNode {
-	workDir, _ := ioutil.TempDir("", "funnel-test-storage-")
-	conf.Worker.WorkDir = workDir
-	log := logger.NewLogger("test-node", logger.DebugConfig())
-
-	// A mock scheduler client allows this code to fake/control the worker's
-	// communication with a scheduler service.
-	res, _ := detectResources(conf.Node, conf.Worker.WorkDir)
-	s := new(MockClient)
-	n := &NodeProcess{
-		conf:      conf,
-		client:    s,
-		log:       log,
-		resources: res,
-		workerRun: NoopWorker,
-		workers:   newRunSet(),
-		timeout:   util.NewIdleTimeout(time.Duration(conf.Node.Timeout)),
-		state:     NodeState_ALIVE,
-	}
-
-	s.On("PutNode", mock.Anything, mock.Anything).
-		Return(nil, nil)
-	s.On("PutNode", mock.Anything, mock.Anything, mock.Anything, mock.Anything).
-		Return(nil, nil)
-	s.On("Close").Return(nil)
-
-	return testNode{
-		NodeProcess: n,
-		Client:      s,
-		done:        make(chan struct{}),
-	}
+type testNode struct {
+	*NodeProcess
+	conn *grpc.ClientConn
+	done chan struct{}
 }
 
 func (t *testNode) Start() context.CancelFunc {
+	t.done = make(chan struct{})
 	ctx, cancel := context.WithCancel(context.Background())
-	t.Client.On("GetNode", mock.Anything, mock.Anything, mock.Anything).
-		Return(&Node{}, nil)
 	go func() {
-		t.NodeProcess.Run(ctx)
+		err := t.NodeProcess.Run(ctx)
+		if err != nil {
+			panic(err)
+		}
 		close(t.done)
 	}()
 	return cancel
@@ -67,28 +58,87 @@ func (t *testNode) Wait() {
 	<-t.done
 }
 
-func (t *testNode) AddTasks(ids ...string) {
-	// Set up the scheduler mock to assign a task to the worker.
-	t.Client.On("GetNode", mock.Anything, mock.Anything, mock.Anything).
-		Return(&Node{
-			TaskIds: ids,
-		}, nil).
-		Once()
+func newTestNode(conf config.Config) *testNode {
+	log := logger.NewLogger("test-node", logger.DebugConfig())
 
-	t.Client.On("GetNode", mock.Anything, mock.Anything, mock.Anything).
-		Return(&Node{}, nil)
+	// Create client for scheduler RPC
+	ctx := context.Background()
+	conn, err := rpc.Dial(ctx, conf.Server)
+	if err != nil {
+		panic(fmt.Errorf("connecting to server: %s", err))
+	}
+	client := NewSchedulerServiceClient(conn)
+
+	node, err := NewNodeProcess(conf.Node, client, NoopWorker, log)
+	if err != nil {
+		panic(err)
+	}
+	return &testNode{NodeProcess: node, conn: conn}
+}
+
+func newTestSched(conf config.Config) *Scheduler {
+	log := logger.NewLogger("test-sched", logger.DebugConfig())
+
+	// Open TCP connection for RPC
+	lis, err := net.Listen("tcp", conf.Server.RPCAddress())
+	if err != nil {
+		panic(err)
+	}
+
+	grpcServer := grpc.NewServer()
+	ev := events.Noop{}
+	sched, err := NewScheduler(conf.Scheduler, log, ev)
+	if err != nil {
+		panic(err)
+	}
+
+	RegisterSchedulerServiceServer(grpcServer, sched)
+	if err != nil {
+		panic(err)
+	}
+
+	go func() {
+		err := grpcServer.Serve(lis)
+		if err != nil {
+			panic(err)
+		}
+	}()
+
+	return sched
 }
 
 func timeLimit(t *testing.T, d time.Duration) func() {
 	stop := make(chan struct{})
+
 	go func() {
 		select {
-		case <-time.NewTimer(d).C:
-			t.Fatal("time limit expired")
+		case <-time.After(d):
+			panic("time limit expired")
 		case <-stop:
 		}
 	}()
+
 	return func() {
 		close(stop)
+	}
+}
+
+func simpleNode() *Node {
+	return &Node{
+		// This ID MUST match the ID set in setup()
+		// because the local scheduler is built to have only a single node
+		Id: "test-node-id",
+		Resources: &Resources{
+			Cpus:   1.0,
+			RamGb:  1.0,
+			DiskGb: 1.0,
+		},
+		Available: &Resources{
+			Cpus:   1.0,
+			RamGb:  1.0,
+			DiskGb: 1.0,
+		},
+		Zone:  "ok-zone",
+		State: NodeState_ALIVE,
 	}
 }
