+++ tests/perf/perf_test.go
@@ -15,7 +15,7 @@ func BenchmarkRunSerialNoNodes(b *testing.B) {
 	fun := tests.NewFunnel(tests.DefaultConfig())
 	defer fun.Cleanup()
 	// No nodes connected in this test
-	fun.Conf.Compute = "manual"
+	fun.Conf.Compute = "builtin"
 	fun.Conf.Logger.Level = "error"
 
 	fun.StartServer()
@@ -32,7 +32,7 @@ func BenchmarkRunConcurrentNoNodes(b *testing.B) {
 	fun := tests.NewFunnel(tests.DefaultConfig())
 	defer fun.Cleanup()
 	// No nodes connected in this test
-	fun.Conf.Compute = "manual"
+	fun.Conf.Compute = "builtin"
 	fun.Conf.Logger.Level = "error"
 	fun.StartServer()
 	b.ResetTimer()
